Reset Functionality for User Previews
It's possible that a user may change their mind and would like to enter another
user name after they've made a submission, a reset button would certainly be
useful in this situation.

The reset functionality has to come from Battle because instances maintain their
own state. Therefore, Battle will now have a reset handler that it can pass down
to other components; the handler will allow components to communicate to a
Battle instance that it must reset it's state. We also pass it an id so that it
can identify which user in the state should be reset.

The UserPreview component now accepts an id and a reset handler that it will use
to reset Battle's state. Since we're invoking a handler with a button, it's not
enough to just assign the handler to onClick, we also have to make sure that the
handler is invoked with the id as an argument. This is why we bind the handler
to the id that passed into UserPreview. The id tells Battle which user s info
should be reset.

app/components/Battle.js
        - Adding the reset functionality to User Previews.
                - To support the reset functionality, Battle now has a reset
                  handler that will allow components to communicate to it that a
                  change of state is necessary. We pass this handler to other
                  components through props.

app/index.css:
        - Adding styling to reset functionality

        - Adding flexbox and centering to the UserPreview Component