Appending a Query String to the Results Path
The idea is to populate the Results page with the stats of both battle users.
This requires us to know the user names within the Results component; we'll be
making a fetch request to the GitHub API for the stats of both users. You might
be thinking at this point "Oh! I know! We'll pass them as props." The trouble
with that is that we want Results to be a component that replaces Battle when a
user hits the Battle button. If we were to pass the user names as props, then
Results would have be rendered as a child of Battle. However, if we do that,
there's no point in assigning a path to Results. Instead, we would have to write
a bunch of conditionals deciding when to wipe all of Battle components and
replace them with the Results component. If Battle has a ton of components, you
can imagine the Battle's render is going to get really messy fast.  To avoid all
of this, we make Results a component that succeeds Battle and we pass it the
necessary information through a query string.

React-Router allows us to do this with the Link Component. Link grants us to the
ability to pass an object that describes the path name that the router is
supposed to route us to, and we can also describe a query string that will be
appended to the pathname. We make use of this functionality to pass the Results
component the necessary data from Battle through the URI's query string.

app/components/Battle.js:
        - The Link Component is passed a query string that will be appended to
          the pathname of the link.
                - We pass both of the submitted user names.