Implementing the Battle Routine and it's Subroutines
We left off with creating a preview component that allows a user to preview
their submission. Users may change their mind after making a submission,
therefore, we allow them to reset their selection. Once the users are happy with
their submissions, they can confirm by pressing "Battle". This will cause the
router to route to a results component, representing a new phase in the battle
process. Since we have the results component render as a separate component, and
not a child of battle,  it doesn't have access to battle's state. Therefore, we
pass the appropriate parameters, through the uri, as query string. React-router
allows us to access a match object, an object created describing the properties
(url, query, etc.) of a matched route, which helps us to greatly in composing a
uri for the results component.

The goal is to create a results component that is able to render the statistics
of the two submitted users, along with their profile name and avatar. In
addition, we also want to calculate a winner and a score based on these
stats. In order to accomplish this, the component will require the means to
fetch the data and calculate a score to determine the winner. This commit
contains a collection of functions designed to do exactly that.

The functions added to the api_helpers module are Promise-based functions; they
allow us to fetch and manipulate data into a usable form for our components
without creating 'callback hell'. We, again, make use of axios to create async
http get requests to fetch data from the GitHub API.

In order to create this battle function that promises, on completion, to pass
over an array of user objects, complete with their score and statistics, we need
to compose it out of smaller functions.

Note: Be sure to read this excellent article on Promises:
https://developers.google.com/web/fundamentals/getting-started/primers/promises

Most of these smaller functions are kept private, in the sense that they are
never exported, and are composed into a larger function called getUserInfo.
getUserInfo, with the help of these smaller functions, fetches the user's
profile and repos from GitHub and crafts a new object containing the user's
profile and the score computed from the repos.

With the expectation that getUserInfo will fetch the user's info, battle invokes
the function for each of the user names it receives and promises, on completion,
to pass over a sorted array -- in descending order, so that we can pick out the
winner by referencing the first element in the array. Since battle immediately
returns with a Promise object -- recall that Promises are async structures and,
therefore, return immediately after invocation/creation --, when we invoke
battle, we'll need to tack on a callback for when the promise completes.

As you can guess, we're going to export and use battle in the Results component
to fetch the data we need. We'll pass battle the submitted user names extracted
from the uri's query string, as an array.

app/utils/api_helpers.js:
        - Defining a bunch of helper functions that fetch data from GitHub using
          the GitHub API
                - These helper functions are Promise-based, and because battle is a
                  composition of these Promise-based functions, battle must also
                  return a promise.

        - Adding helper functions that sum up the number of stars each user in
          their repos.