Implementing the Results Component's Functionality
Up to this point we have everything we need to implement the Results Component.
But before the UI is built for the Results Component, we'll need to fetch data
from the GitHub API, and we'll need to organize it before styling it.

First, we need to answer a few questions: Does this component have any state? If
so, what is the state, and what kind of information does it need to keep track of?

This component will move from state to state, so, what do we need to keep track
of?  We'll need to track both users, whether a component is loading, and whether
it has failed. The state moves from being empty to loading to either filled with
data or an error message, in the case where it fails. The properties that
describe the "filled with data" state is winner, loser, and loading. The
properties that define an error state is error and loading (if an error occurs,
then the component is no longer loading). The loading state is determined by the
loading property. These four properties describe the state the component is in.

We employ help from battle in our API helper module to fetch data using the
GitHub API. We also make use of the query-string module to parse the query
string in the url, which contains the usernames necessary to make a call to
battle. Thankfully, react-router passes the location object as a prop; we can
grab the query string and parse it to get our usernames.

One question arouse while I was initially building this, and it almost stumped
me. How do we move data out of a Promise chain. Is it possible or do we just
continue adding to it? A Promise allows you to chain together callbacks where it
promises to pass the results over once the Promise has resolved, that is, once
the Promise's callback has executed. Tacking on a .then or .catch creates a new
Promise which you can tack on another .then or.catch with another callback. So
if we have a string of these callbacks, how do we get the data out of this
chain. The answer is we use closures to our advantage and have the callback
reference an outer variable and assign the data to it. Similarly, this is what
do with our APIHelpers. In a callback, we move the data out of the chain by
making a reference to the Component's setState (the state is a structure outside
the Promise chain) and store the data in the state.

Just a quick note: The error handling here is a bit different from what I've
encountered. When an error occurs in battle, it is caught, outputs the error
message, and returns null. I wasn't sure how to deal with the null. It doesn't
give any information about the error. Therefore, we create our own message to
display to the user in the component. So when there is an error, change the
error property to an error message describing the error. We'll use the fact that
this property is "truthy" to check whether an error occurs or not. We can then
immediately return a UI from render. It's basically the "fail-fast" approach
most of us are familiar with. It's just that I've never used it in React. I
hope that this note will be helpful to others that feel a bit tangled.

app/component/Results.js:
        - Fetch data from the GitHub API using the our APIHelper module. Since
          it is Promised-based, we tack on a callback in order to access the
          data in the Promise. Within the Promise, we store the fetched data by
          storing it in the state of the component.

        - Player has been added as a stateless component. It's responsibilities
          are just to display the data passed into it. Results encapsulates it
          and maintains the state of the entire component.

        - The UI has yet to be built. As the moment, a rough draft of how the
          data would be organized has been written. Most of the work done was
          getting data from the GitHub API and maintaining the state of the
          component. Next, we'll continue to organize the fetched data. Once
          the markup has been written, then we'll start to style it.