Making UserPreview More Dynamic (Generalizing the Component)
While organizing the data retrieved from the fetch, writing the username and
avatar portions seems extremely familiar. Of course, it was almost exactly like
UserPreview component that was in the Battle component. The only difference is
that we hard-coded a reset button into the UserPreview component, something that
wasn't needed in the Results component. But before perform this generalization,
we need to recognize that UserPreview has to be it's own module if it is be
reusable. UserPreview is removed from Battle as a private component and placed
into is own file where it is exported as a module.

Now, in order to reconcile both use cases, the UserPreview component is
generalized to where it's composed with its children, that is, it is generalized
to render its children. Both the username and the avatar are integral to the
component, but the reset button is not. So we made it so that we can pass any
element or component as a child and expect that it will be render with an avatar
and a username. This adds an element of dynamism to the component. We are no
longer restricted in its use because of the hard-coded reset.

The UI still hasn't been realized because the content is still being organized and
marked up, style comes later.

/app/components/Battle.js:
        - Moving the UserPreview component out of Battle into its own file.
          UserPreview is now its own component.

        - Since UserPreview is accepts and renders children, we pass the reset
          button here because its specific to the Battle Component.
                - Also, since we have access to the handlers directly (because
                  the button is no longer encapsulated in UserPreview), we
                  modify what gets passed into the button. Before, when it was
                  embedded in UserPreview, we would have to pass the handlers
                  into UserPreview before it could be used by the button.
                  This required a ton of binding to get this working correctly.
                  Now, we can reference the handler directly, though, some binding
                  may be required.

        - A mistake has been correct with the Link URL for the confirmation
          button. 'results/' was being appended, but it's wrong because the
          query would be appended immediately afterward creating the url
          "...battle/results/?querystring".

/app/components/Results.js:
        - We make use of the new UserPreview component to render a user's avatar
          and username.

/app/components/UserPreview.js:
        - We remove the reset button from the component and instead render its
          children. This allows a user of the component to define what should be
          rendered with a username and password. Mainly, we want UserPreview to
          be reusable for both Battle and the Player component in Results.