Highly Reusuable Containers (Also Adding PropTypes)
The Loading component was created to be a super reusable component, that is,
it's suppose to be component where it doesn't depend on other components and
it's styles are mostly contained (we inline styles instead of depending on an
external stylesheet). The nice part of creating these highly reusable components
is that they can be uploaded to npm and can be used as part of other projects.
The lack of dependencies on this app makes this possible.

For our purposes, we use Loading anywhere an async load occurs.

I'll comment on a few issues that I've encountered while building this.

First, I kept getting a console warning stating that setState was being called
on an unmounted DOM. It occurred to me that I needed to end the interval, but
somehow I convinced myself that once Loading was unmounted the interval would
end. Turns out that is no the case at all, the callback set with the interval
function kept being invoked after it was unmounted. The solution to this was
clear the interval once the component was unmounted.

I also found out that class components have properties that allow
functions to be invoked. For instance, the .interval property allows for
the object that setInterval returns. And it makes sense, because in order to
kill this process that continuously invokes its callback, we have to call
clearInterval on it. clearInterval doesn't keep track of all the routines
running, it needs to be called on the object itself. What I'm not sure about is
that is 'interval' a reserved property or just a regular property that we tacked
on; I suspect the latter.

Initially, I built so that the decision to reset the Loading message (after the
third dot was appended) was done inside of setState. The check would determine
what to return to update the state. As it turns out, it required two binds and that
might confuse users, as opposed to calling setState in two different places and
making one one bind.

As a final thing, we'll add PropTypes to this component in the next commit.

/app/components/Results.js:
        - Adding PropTypes to some of the smaller components

        - Replacing the loading message with a Loading component

/app/components/Popular.js:
        - Replacing the loading message with a Loading component

/app/components/Loading.js:
        - A highly reusable component whose style is self-contained and has no
          dependencies to this app. It can be used anywhere, but it's intended
          to be used anywhere an async request occurs.